Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.9.2
pandas==2.1.3
scikit-learn==1.6.1
joblib==1.3.2
torch==2.6.0
torchvision==0.21.0
albumentations==2.0.5
Pillow==10.1.0
firebase-admin==6.2.0
python-multipart==0.0.6 